QAeonCoreDevelopment/pyproject.toml
[project]
name = "quantum-aeon-fluxor"
version = "0.1.0"
description = "An advanced AI research project exploring consciousness and complex systems."
requires-python = ">=3.10"
dependencies = [
    "google-generativeai>=0.4.0",
    "python-dotenv>=1.0",
    "pydantic>=2.6",
    "qdrant-client>=1.15.1",
]

[project.optional-dependencies]
dev = [
    "pytest",
    "ruff",
    "black",
]

[project.scripts]
qaf-cli = "quantum_aeon_fluxor.main:main"
qaf-index = "quantum_aeon_fluxor.hermetic_engine__persistent_data.indexing.index_folder:index_folder"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv]
dev-dependencies = [
    "pytest",
    "ruff",
    "black",
]

[tool.ruff]
line-length = 100

[tool.black]
line-length = 100
